chore: update on home
