+++ src/pages/Home/index.tsx
@@ -1,7 +1,20 @@
-import { FC } from 'react'
+import { FC, useState } from 'react'
 
-import { Text, CardConfig, Card, CardResult } from 'components'
-import { Colors, FontFamilies, FontSizes, Icons } from 'shared/style/constants'
+import {
+  Text,
+  CardConfig,
+  Button,
+  LineChart,
+  Card,
+  CardResult,
+} from 'components'
+import {
+  Colors,
+  ButtonAppearance,
+  FontFamilies,
+  FontSizes,
+  Icons,
+} from 'shared/style/constants'
 import { MainLayout } from 'layouts/MainLayout'
 import { useGetReportsByIdQuery, useGetStatByIdQuery } from 'shared/store/apis'
 import { useAppSelector } from 'shared/store'
@@ -13,6 +26,7 @@ import { StyledHome } from './core/home.styles'
 
 export const Home: FC = () => {
   const selectedRunId = useAppSelector((state) => state.general.selectedRunId)
+  const [showCumulativeChart, setShowCumulativeChart] = useState(false)
 
   useHome({ selectedRunId })
 
@@ -22,6 +36,27 @@ export const Home: FC = () => {
   const { data: report, isLoading: isReportLoading } = useGetReportsByIdQuery(
     selectedRunId ?? skipToken
   )
+
+  const graphData = report?.benchmark.records.map((event, i) => {
+    return { x: i, y: convertInMs(event.time) }
+  })
+
+  const graphDataCumulative = report?.benchmark.records.map((event, i) => {
+    if (i > 0) {
+      let time = 0
+      for (let j = 0; j <= i; j++) {
+        time += report?.benchmark.records[j].time
+      }
+      return {
+        x: i,
+        y: convertInMs(time),
+      }
+    }
+    return {
+      x: i,
+      y: convertInMs(event.time),
+    }
+  })
 
   return (
     <MainLayout>
@@ -48,7 +83,44 @@ export const Home: FC = () => {
               Results
             </Text>
             <div className="f f-jc-space-b mt-2">
-              <Card className="chart mr-2">charts</Card>
+              <Card className="chart mr-2">
+                <div className="space-between">
+                  <Text
+                    className="mb-4"
+                    font={FontFamilies.POPPINS}
+                    size={FontSizes.LARGE}
+                  >
+                    Graphs
+                  </Text>
+                  <Button
+                    appearance={ButtonAppearance.OUTLINE}
+                    mainColor={Colors.WHITE}
+                    secondaryColor={Colors.BLACK}
+                    hover={true}
+                    active={true}
+                    onClick={() =>
+                      showCumulativeChart
+                        ? setShowCumulativeChart(false)
+                        : setShowCumulativeChart(true)
+                    }
+                  >
+                    {showCumulativeChart
+                      ? 'Show graph per request'
+                      : 'Show cumulative graph per request'}
+                  </Button>
+                </div>
+                {showCumulativeChart ? (
+                  <LineChart
+                    legendY="time/req"
+                    data={[{ id: 'linear', data: graphData }]}
+                  />
+                ) : (
+                  <LineChart
+                    legendY="time/req"
+                    data={[{ id: 'cumulative', data: graphDataCumulative }]}
+                  />
+                )}
+              </Card>
               <div className="stats">
                 <CardResult
                   icon={Icons.CLOCK}
